requirements.txt
langgraph
langgraph-sdk
langgraph-checkpoint-sqlite
langsmith
langchain-community
langchain-core
langchain-openai
tavily-python
wikipedia
trustcall
langgraph-cli
python-dotenv
streamlit
langserve
fastapi
uvicorn
sse_starlette
bs4
pypdf
chromadb
faiss-cpu
requests
beautifulsoup4
groq
langchain-groq